requirements.txt
pandas==2.2.0
google-genai==0.3.0
python-dotenv==1.0.1
numpy==1.26.4